pyproject.toml
[project]
name = "easycoffee-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Rasul Akhmetov",email = "rasulakhmetov2004@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "django-ninja (>=1.4.3,<2.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "django-ninja-extra (>=0.30.1,<0.31.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "injector (>=0.22.0,<0.23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
